+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Properly set the dimensions of pseudo-terminals on Linux
 - Ensure that CLI formatting utilizes the full width of the terminal
+- Don't use pseudo-terminals when running in CI
 
 ## 0.23.1 - 2025-07-25