+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
+## 0.24.0 - 2025-08-06
+
 ***Added:***
 
 - Add `build bazel` command